:recycle: Split code paths for assignment exprs
We can split the code path into assignments to members, and assignments
in other cases, avoiding the need to make some checks when the LHS is
not a MemberExpressionAst.

Reworked things a bit to use Lazy<string> for event definition.
Also refactored logic to use StringBuilder completely per Rob's comments

Co-authored-by: Robert Holt <rjmholt@gmail.com>